chore: use BigDecimal for price input, rename AuctionStatus OPEN to ACTIVE, move auction application code to feature package

+++ auction-service/src/main/java/wpessers/auctionservice/application/port/in/command/CreateAuctionCommand.java
@@ -1,5 +1,6 @@
 package wpessers.auctionservice.application.port.in.command;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 
 public record CreateAuctionCommand(
@@ -7,7 +8,7 @@ public record CreateAuctionCommand(
     String description,
     Instant startTime,
     Instant endTime,
-    double startingPrice
+    BigDecimal startingPrice
 ) {
 
 }
